+++ docs/changes.rst
@@ -17,8 +17,7 @@ Release History
 * messages are now iterable. You can do ``for byte in msg:`` or
   ``list(msg)``. (Not implemented for meta messages.)
 
-* messages now have an ``is_meta`` attribute which can be used instead
-  of ``isinstance(msg, MetaMessage)``.
+* new message attributes ``is_meta`` and ``is_realtime``.
 
 * added `FrozenMessage` and `FrozenMetaMessage`. These are immutable
   versions of `Message` and `MetaMessage` that are hashable and thus